requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
pdfminer.six
python-docx
pytesseract
opencv-python
python-multipart
